Refactory and fix bugs
    update cpp interface
    Separate emi_sock to emi_sockr, so the send only lib only use std lib
    Clarify return value for block mode
    Find and fix shbuf access NULL and mem free bugs
    hlist replacement of msg_map done
    Fix segmentfault bug, thread miss set args as NULL with pool exhausted
    basic release on exit done
    Add -Wextra and fix warnings
    fix multi release memory problem Need to further fix the atomic problem: read(check) then release
    combine hlist delete and free msg_map memory into one for locking
    Change emi_semaphore to emi_lock
    Fix emi_receive_operation args double free problem.  don't need to alloc args for emi_receive_operation

+++ src/emi_core.c
@@ -16,42 +16,28 @@
 #include <pthread.h>
 
 #include "emi_msg.h"
-#include "msg_table.h"
+#include "emi_hash.h"
 #include "emi_shbuf.h"
-#include "emi_sock.h"
-#include "emi_semaphore.h"
+#include "emi_sockr.h"
+#include "emi_lock.h"
 #include "emi_dbg.h"
 #include "emi_config.h"
 #include "emi_shmem.h"
 #include "emi_thread.h"
 
 
-struct clone_args{
-    struct sk_dpr *sd;
-    struct sk_dpr *client_sd;
-};
-
 static int emi_receive_operation(void *args);
 
 
 static int core_shmid=-1;
-static struct sk_dpr *sd=NULL;
-static struct sk_dpr *client_sd=NULL;
-
 static struct emi_shmem_mgr core_shmem_mgr;
-
-static struct msg_map *msg_table[EMI_MSG_TABLE_SIZE];
-
 static struct emi_thread_pool emi_core_pool;
 
 void emi_release(void){
-    emi_close(sd);
-    emi_close(client_sd);
     if(core_shmid>=0){
         emi_shm_destroy("emilib", core_shmid);
     }
 }
-
 
 static int int_global_shm_space(int pid_max){
     void *base;
@@ -102,8 +88,9 @@ int emi_core(struct emi_config *config){
 static int __emi_core(void){
 
     eu32 pid_max;
+    struct sk_dpr *sd=NULL;
 
-    if(init_msg_table_lock(msg_table)){
+    if(init_msg_table_lock(msg_table_head, &msg_table_lock)){
         emilog(EMI_ERROR, "init msg table error\n");
         return -1;
     }
@@ -150,42 +137,19 @@ static int __emi_core(void){
     }
 
     while(1){
+        struct sk_dpr *client_sd;
+
         if((client_sd=emi_accept(sd,NULL))==NULL){
             emilog(EMI_WARNING, "emi_accept error\n");
-            emi_close(client_sd);
             continue;
         }
 
-        struct clone_args *arg;
+        emi_thread_pool_submit(&emi_core_pool, (void *)emi_receive_operation, (void *)client_sd);
 
-        if((arg=(struct clone_args *)malloc(sizeof(struct clone_args)))==NULL){
-            emilog(EMI_WARNING, "mem error\n");
-            continue;
-        }
-
-        arg->client_sd=client_sd;
-
-        emi_thread_pool_submit(&emi_core_pool, (void *)emi_receive_operation, arg);
-
-
-   /* 
-        pthread_t tid;
-
-        if(pthread_create(&tid,NULL,(void *)emi_receive_operation,arg)){
-            emilog(EMI_WARNING, "pthread cancel error\n");
-            continue;
-        }
-
-        if(pthread_detach(tid)){
-            emilog(EMI_WARNING, "pthread_detach error\n");
-            continue;
-        }
-        */
-    
     }
 }
 
-static int emi_receive_operation(void *args){
+static int emi_receive_operation(void *client_sd){
     int ret = -1;
     struct emi_msg *msg_pos;
 
@@ -201,7 +165,7 @@ static int emi_receive_operation(void *args){
  * read the remote msg into this alloced memory, if this one got an error, probably emi_init has sent a guess message to emi_core, which
  * is a connection to emi_core without any data transfered.
  */
-    if((ret=emi_msg_read(((struct clone_args *)args)->client_sd, msg_pos)) < 0){
+    if((ret=emi_msg_read((struct sk_dpr *)client_sd, msg_pos)) < 0){
         emilog(EMI_INFO, "emi_read from client error or emi_init is guessing port\n");
         goto e0;
     }
@@ -210,23 +174,28 @@ static int emi_receive_operation(void *args){
     if(msg_pos->flag&EMI_MSG_CMD_REGISTER){
         emilog(EMI_DEBUG, "Received a register msg with num %d and pid %d\n",
                 msg_pos->msg, msg_pos->addr.pid);
+        
+        // FIXME
+        struct msg_map *p = alloc_msg_map();
+        if(p == NULL){
+            goto e0;
+        }
 
-        struct msg_map p;
-        msg_map_init(&p,msg_pos->msg,msg_pos->addr.pid);
+        msg_map_init(p,msg_pos->msg,msg_pos->addr.pid);
 
         /*
          * Make sure msg_map with the same pid and msgnum are not in msg_table,
          * this prevents one process register multiple functions to the same msg number.
          */
-        if((ret = emi_hinsert_lock(msg_table,&p)) < 0){
+        if((ret = emi_hinsert_unique_lock(msg_table_head, p)) < 0){
             msg_pos->flag &= ~EMI_MSG_RET_SUCCEEDED;
         }else{
-            emi_lock_init(&core_shmem_mgr.pididx_lock[p.pid]);
+            emi_lock_init(&core_shmem_mgr.pididx_lock[p->pid]);
             msg_pos->flag |= EMI_MSG_RET_SUCCEEDED;
         }
 
         emilog(EMI_DEBUG, "Insert msg to msg_table done");
-        if((ret=emi_msg_write_payload(((struct clone_args *)args)->client_sd,msg_pos)) < 0){
+        if((ret=emi_msg_write_payload((struct sk_dpr *)client_sd, msg_pos)) < 0){
             emilog(EMI_WARNING, "emi read payload from client error\n");
         }
 
@@ -236,72 +205,71 @@ static int emi_receive_operation(void *args){
 
     }else{
 
-        struct msg_map p;
         int nth;
         eu32 *pid_num_addr;
-        int num;
-        struct msg_map *map;
 
         emilog(EMI_DEBUG, "Received a sending msg with num %d, data size = %d\n", msg_pos->msg, msg_pos->size);
         debug_emi_msg(msg_pos);
 
         nth=get_shbuf_offset(core_shmem_mgr.base, msg_pos);
 
-        msg_map_init(&p,msg_pos->msg,0);
+        // set fail by default
+        msg_pos->flag&=~EMI_MSG_RET_SUCCEEDED;
 
-        //msg_pos->flag&=~EMI_MSG_RET_SUCCEEDED;
-        msg_pos->flag |= EMI_MSG_RET_SUCCEEDED;
+        struct hlist_node *tmp;
+        struct msg_map *map;
 
-        for(num=0;;num+=1){
-            emi_spin_lock(&msg_table_lock);
-            if((map=__emi_hsearch(msg_table,&p,&num))==NULL){
-                emi_spin_unlock(&msg_table_lock);
-
-                //if no registered msg in msg_table, return failed
-                if(num == 0)
-                    msg_pos->flag&=~EMI_MSG_RET_SUCCEEDED;
-                break;
-            }
-            emi_spin_unlock(&msg_table_lock);
-
-            pid_num_addr = &core_shmem_mgr.pididx[map->pid];
-            emi_lock(&core_shmem_mgr.pididx_lock[map->pid]);
-            emilog(EMI_DEBUG, "pid %d found, for msg %d, cmd %d\n", map->pid, msg_pos->msg, msg_pos->cmd);
-
-            *pid_num_addr = nth;
-            int pid_ret=kill(map->pid,SIGUSR2);
-            if(pid_ret<0){
-                *pid_num_addr=0;
-                //emi_spin_lock(&msg_pos->lock);
-                //msg_pos->flag&=~EMI_MSG_RET_SUCCEEDED;
-                //emi_spin_unlock(&msg_pos->lock);
-                emi_hdelete_lock(msg_table,map);
-
-            }else{
-                emi_spin_lock(&msg_pos->lock);
-                emilog(EMI_DEBUG, "msg_pos->count = %d", msg_pos->count);
+        emi_hsearch(msg_table_head, map, node, msg_pos->msg){
+            if(map->msg == msg_pos->msg){
                 msg_pos->count++;
-                emi_spin_unlock(&msg_pos->lock);
             }
-            emilog(EMI_DEBUG, "kill return %d for pid %d\n", pid_ret, map->pid);
+        }
 
-            while (*pid_num_addr) {
-                sched_yield(); //Need discussion
-            };
+        emi_hsearch(msg_table_head, map, node, msg_pos->msg){
+            if(map->msg == msg_pos->msg){
+                pid_num_addr = &core_shmem_mgr.pididx[map->pid];
+                emi_lock(&core_shmem_mgr.pididx_lock[map->pid]);
+                emilog(EMI_DEBUG, "pid %d found, for msg %d, cmd %d\n", map->pid, msg_pos->msg, msg_pos->cmd);
 
-            emi_unlock(&core_shmem_mgr.pididx_lock[map->pid]);
+                *pid_num_addr = nth;
+                int pid_ret=kill(map->pid,SIGUSR2);
+
+                if(pid_ret<0){
+                    *pid_num_addr=0;
+                }
+
+                emilog(EMI_DEBUG, "kill return %d for pid %d\n", pid_ret, map->pid);
+               
+                while (*pid_num_addr) {
+                    sched_yield(); //Need discussion
+                };
+
+                emi_unlock(&core_shmem_mgr.pididx_lock[map->pid]);
+            }
         }
 
         emilog(EMI_DEBUG, "Signal all processes with msg num %d, waiting for sync\n", msg_pos->msg);
+
         while(msg_pos->count){
-            sched_yield();
+            emi_hsearch_safe(msg_table_head, map, tmp, node, msg_pos->msg){
+                if(map->msg == msg_pos->msg){
+                    int pid_ret=kill(map->pid, 0);
+                    if(pid_ret<0){
+                        emi_spin_lock(&msg_pos->lock);
+                        msg_pos->count--;
+                        emi_spin_unlock(&msg_pos->lock);
+
+                        free_msg_map(map);
+                    }
+                }
+            }
         }
 
         emilog(EMI_DEBUG, "msg_pos->flag = %x\n", msg_pos->flag);
         if(msg_pos->flag&EMI_MSG_MODE_BLOCK){
             emilog(EMI_DEBUG, "Return the state and possible data for block mode\n");
 
-            if((ret = emi_msg_write_ret(((struct clone_args *) args)->client_sd, msg_pos))< 0){
+            if((ret = emi_msg_write_ret((struct sk_dpr *)client_sd, msg_pos))< 0){
                 goto e0;
             }
 
@@ -314,9 +282,6 @@ static int emi_receive_operation(void *args){
 e0:
     free_shared_msg(msg_pos);
 e1:
-    emi_close(((struct clone_args *)args)->client_sd);
-    free(args);
-    //pthread_exit(NULL);
+    emi_close((struct sk_dpr *)client_sd);
     return ret;
-
 }
